+++ src/main/java/com/clubber/ClubberServer/global/infrastructure/outer/api/oauth/client/kakao/KakaoInfoClient.java
@@ -2,7 +2,7 @@ package com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.client.k
 
 import com.clubber.ClubberServer.domain.auth.dto.UnlinkKaKaoTarget;
 import com.clubber.ClubberServer.global.config.feign.FeignConfig;
-import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.dto.kakao.KakaoUserInfoResponse;
+import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.kakao.dto.KakaoUserInfoResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.PostMapping;
