+++ src/main/java/com/clubber/ClubberServer/global/infrastructure/outer/api/oauth/client/kakao/KakaoOauthClient.java
@@ -1,7 +1,7 @@
 package com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.client.kakao;
 
 import com.clubber.ClubberServer.global.config.feign.FeignConfig;
-import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.dto.kakao.KakaoTokenResponse;
+import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.kakao.dto.KakaoTokenResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
